Moved functions clamp, inRange, mod, $, $$ into utils.js
